Fixed two GCC POSIX warnings
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@6849 212acab6-be3b-0410-9dea-997c60f758d6

+++ Contrib/Library/LibraryLocal/LibraryLocal.cpp
@@ -150,8 +150,8 @@ int _tmain(int argc, TCHAR* argv[])
   }
   else
   {
-    fprintf(fHdr, "!define LIBRARY_VERSION_HIGH %lu\n", high);
-    fprintf(fHdr, "!define LIBRARY_VERSION_LOW %lu\n", low);
+    fprintf(fHdr, "!define LIBRARY_VERSION_HIGH %lu\n", static_cast<unsigned long>(high));
+    fprintf(fHdr, "!define LIBRARY_VERSION_LOW %lu\n", static_cast<unsigned long>(low));
   }
 
   fclose(fHdr);
